+++ src/dmx/hal/gpio.h
@@ -8,9 +8,9 @@
  */
 #pragma once
 
-#include "dmx_types.h"
+#include "dmx/types.h"
 #include "driver/gpio.h"
-#include "rdm_types.h"
+#include "rdm/types.h"
 
 #ifdef __cplusplus
 extern "C" {
