+++ src/esp_dmx.h
@@ -12,8 +12,8 @@
 #include "dmx/device.h"
 #include "dmx/driver.h"
 #include "dmx/rw.h"
-#include "dmx_types.h"
-#include "rdm_types.h"
+#include "dmx/types.h"
+#include "rdm/types.h"
 
 #ifdef __cplusplus
 extern "C" {
