+++ src/rdm/controller.h
@@ -8,8 +8,8 @@
 
 #include <stdint.h>
 
-#include "dmx_types.h"
-#include "rdm_types.h"
+#include "dmx/types.h"
+#include "rdm/types.h"
 
 #ifdef __cplusplus
 extern "C" {
