+++ src/rdm/responder.h
@@ -8,9 +8,9 @@
 
 #include <stdint.h>
 
-#include "dmx_types.h"
+#include "dmx/types.h"
 #include "rdm/parameters.h"
-#include "rdm_types.h"
+#include "rdm/types.h"
 
 #ifdef __cplusplus
 extern "C" {
