+++ src/rdm_utils.h
@@ -14,9 +14,9 @@
 #include <stdint.h>
 #include <string.h>
 
-#include "dmx_types.h"
+#include "dmx/types.h"
 #include "rdm/responder.h"
-#include "rdm_types.h"
+#include "rdm/types.h"
 
 #ifdef __cplusplus
 extern "C" {
